Some tweaks and changes.
