+++ Assets/Animations.meta
@@ -1,0 +1,9 @@
+fileFormatVersion: 2
+guid: 7cb76659df53915439bf2370bfb7676b
+folderAsset: yes
+timeCreated: 1473711995
+licenseType: Free
+DefaultImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
