+++ Assets/Scenes/TestScene.unity.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 8299756d61da2aa4580116ea410a43e3
+timeCreated: 1473711811
+licenseType: Free
+DefaultImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
