+++ Assets/Scripts/ChildRemover.cs
