+++ Assets/Scripts/FearInfo.cs
