+++ Assets/Scripts/FlashlightController.cs
