+++ ProjectSettings/ProjectVersion.txt
@@ -1,2 +1,2 @@
-m_EditorVersion: 5.3.4f1
+m_EditorVersion: 5.3.5f1
 m_StandardAssetsVersion: 0
